api/channels/{channelId}/autosharing/facebook DELETE POST GET operation rewrite
git-svn-id: https://everest.teltel.com:8443/repos/misc/9x9dev/trunk/nncloudtv@4115 a0a5f996-2bb4-401c-8778-395c3c01700f

+++ src/main/java/com/nncloudtv/service/NnChannelPrefManager.java
@@ -25,15 +25,53 @@ public class NnChannelPrefManager {
 		return prefDao.save(pref);
 	}
 	
+	public List<NnChannelPref> save(List<NnChannelPref> prefs) {
+        Date now = new Date();
+        for (NnChannelPref pref : prefs) {
+            if (pref.getCreateDate() == null) {
+                pref.setCreateDate(now);
+            }
+            pref.setUpdateDate(now);
+        }
+        return prefDao.saveAll(prefs);
+    }
+	
 	public List<NnChannelPref> findByChannelId(long channelId) {
 		return prefDao.findByChannelId(channelId);
 	}
 	
-	public NnChannelPref findByChannelIdAndItem(long channelId, String item) {
+	public List<NnChannelPref> findByChannelIdAndItem(long channelId, String item) {
 		return prefDao.findByChannelIdAndItem(channelId, item);
 	}
 	
 	public void delete(NnChannelPref pref) {
-		prefDao.delete(pref);
+	    if (pref != null) {
+	        prefDao.delete(pref);
+	    }
+	}
+	
+	public void delete(List<NnChannelPref> prefs) {
+	    if (prefs != null && prefs.size() != 0) {
+	        prefDao.deleteAll(prefs);
+	    }
+    }
+	
+	public String composeFacebookAutoshare(String fbUserId, String accessToken) {
+        return fbUserId + "," + accessToken;
+    }
+	
+	public String[] parseFacebookAutoshare(String value) {
+	    int seperatorIndex = value.indexOf(',');
+	    if(seperatorIndex == -1) {
+	        return null;
+	    } else if (seperatorIndex + 1 == value.length()) { // the ',' is the last one char
+	        return null;
+	    } else if (seperatorIndex == 0) { // the ',' is the first one char
+	        return null;
+	    }
+	    String[] result = new String[2];
+	    result[0] = value.substring(0, seperatorIndex); // facebookID
+	    result[1] = value.substring(seperatorIndex + 1); // accessToken
+	    return result;
 	}
 }
